+++ src/commonMain/kotlin/org/eclipse/keyple/keypleless/distributed/client/protocol/KeypleRemoteService.kt
@@ -11,24 +11,24 @@
  ************************************************************************************** */
 @file:OptIn(ExperimentalStdlibApi::class)
 
-package org.eclipse.keyple.keyplelessdistributedlib.protocol
+package org.eclipse.keyple.keypleless.distributed.client.protocol
 
 import io.github.aakira.napier.Napier
+import kotlin.uuid.ExperimentalUuidApi
+import kotlin.uuid.Uuid
 import kotlinx.serialization.ExperimentalSerializationApi
 import kotlinx.serialization.KSerializer
 import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.jsonObject
-import org.eclipse.keyple.keyplelessdistributedlib.network.KeypleServer
-import org.eclipse.keyple.keyplelessdistributedlib.network.KeypleServerConfig
-import org.eclipse.keyple.keyplelessdistributedlib.network.buildHttpClient
+import org.eclipse.keyple.keypleless.distributed.client.network.KeypleServer
+import org.eclipse.keyple.keypleless.distributed.client.network.KeypleServerConfig
+import org.eclipse.keyple.keypleless.distributed.client.network.buildHttpClient
 import org.eclipse.keyple.keyplelessreaderlib.CardIOException
 import org.eclipse.keyple.keyplelessreaderlib.LocalNfcReader
 import org.eclipse.keyple.keyplelessreaderlib.MultiplatformReader
 import org.eclipse.keyple.keyplelessreaderlib.ReaderIOException
 import org.eclipse.keyple.keyplelessreaderlib.UnexpectedStatusWordException
-import kotlin.uuid.ExperimentalUuidApi
-import kotlin.uuid.Uuid
 
 private const val TAG = "KeypleRemoteService"
